MVC implemented successfully 0.1.2
bodies are now in character class, not in view classes, body height and width are initialized for the first construction. In short, view classes are now just for viewing, physics calculations are in character classes.
