+++ core/src/characters/Character.java
@@ -8,17 +8,64 @@ import helpers.GameInfo;
 
 public class Character {
     
+	private World world;
+	private Body body;
 	private float xPosition;
 	private float yPosition;
 	private float height;
 	private float width;
     
     
-    public Character( float x, float y) {
+    public Character(World world, float initialX, float initialY) {
 
-        setPosition( x , y );
-       
+    	this.world = world;
+    	this.setHeight(10 /* just an initialization */ );
+    	this.setWidth(10 /* just an initialization */ );
+    	
+    	this.setPosition(initialX, initialY);
+    	
+    	createBody();
+    	updateCharacter();
+    	
     }
+    
+    
+    public void createBody() {
+		
+        BodyDef bodyDef = new BodyDef();
+        bodyDef.type = BodyDef.BodyType.DynamicBody;
+        bodyDef.position.set( this.getXPosition() / GameInfo.PPM, this.getYPosition() / GameInfo.PPM );
+        body = world.createBody( bodyDef );
+
+        PolygonShape shape = new PolygonShape();
+        shape.setAsBox( (this.getWidth() / 2f) / GameInfo.PPM,(this.getHeight() / 2f) / GameInfo.PPM);
+
+        FixtureDef fixtureDef = new FixtureDef();
+        fixtureDef.density = 4f; //Mass of the body
+        fixtureDef.friction = 2f; //To not slide on surfaces
+        fixtureDef.shape = shape;
+
+        Fixture fixture = body.createFixture( fixtureDef );
+
+        shape.dispose();
+    
+	}
+
+
+	public void moveCharacter( float x, float y ) {
+		
+	    body.setLinearVelocity( x , y );
+	}
+	
+	public void updateCharacter() {
+		
+        this.setPosition( body.getPosition().x * GameInfo.PPM, body.getPosition().y * GameInfo.PPM);
+    }
+    
+	public Body getBody() {
+		
+		return body;
+	}
 
     public void setPosition(float x, float y) {
 		
@@ -55,10 +102,5 @@ public class Character {
     	
     	return this.width;
     }
-
-
-
-
-
 
 }
